Moved shared preprocessor functions to _shared module
