+++ esmvalcore/preprocessor/_volume.py
@@ -15,11 +15,11 @@ import numpy as np
 from iris.coords import AuxCoord, CellMeasure
 from iris.cube import Cube
 
-from ._area import _try_adding_calculated_cell_area
 from ._shared import (
     get_iris_aggregator,
     get_normalized_cube,
     preserve_float_dtype,
+    try_adding_calculated_cell_area,
     update_weights_kwargs,
 )
 from ._supplementary_vars import register_supplementaries
@@ -155,7 +155,7 @@ def calculate_volume(cube: Cube) -> da.core.Array:
 
     # Get or calculate the horizontal areas of the cube
     has_cell_measure = bool(cube.cell_measures('cell_area'))
-    _try_adding_calculated_cell_area(cube)
+    try_adding_calculated_cell_area(cube)
     area = cube.cell_measure('cell_area').copy()
     area_dim = cube.cell_measure_dims(area)
